Refaktorert til å kunne sikre individuelle ruter
