+++ my-page-app/src/User.js
@@ -1,10 +1,11 @@
 export class User {
-  constructor(name, email, givenName, familyName, icon, startDate) {
+  constructor(name, email, givenName, familyName, icon, startDate, loaded) {
     this.name = name;
     this.email = email;
     this.givenName = givenName;
     this.familyName = familyName;
     this.icon = icon;
     this.startDate = startDate;
+    this.loaded = loaded;
   }
 }
